+++ frontend/src/components/ReportSelect/fields.tsx
@@ -264,9 +264,14 @@ export const reportType = {
     },
     {
       reportCategory: "divisional_project_reports",
-      value: "multi_year_statistics_for_project_change_requests",
+      value: "rpt_PA_ChangeRequestTypesFYSummary",
       label: "Multi-Year Statistics for Project Change Requests",
-      reportParamCategory: [],
+      reportParamCategory: [
+        {
+          field: fieldTypes.fiscal,
+          type: requestTypes.route,
+        },
+      ],
     },
   ],
 };
@@ -349,7 +354,7 @@ export const reportDescription = {
     {
       value:
         "Project Information, Budget Forecasting Information broken down between deliverable, detail amounts, quarter and portfolio recovery amount.",
-      reportType: "multi_year_statistics_for_project_change_requests",
+      reportType: "rpt_PA_ChangeRequestTypesFYSummary",
     },
   ],
 };